requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
oauthlib==3.1.1
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
